moved shared code to common location
